Reducing chatiness
