+++ build.gradle
@@ -53,7 +53,7 @@ subprojects {
     }
 
     ext {
-        javafxVersion = "21.0.1"
+        javafxVersion = "22"
     }
 
 
@@ -79,7 +79,7 @@ subprojects {
             implementation 'jakarta.activation:jakarta.activation-api:2.1.2'
             implementation 'javax.inject:javax.inject:1'
             implementation 'javax.servlet:javax.servlet-api:4.0.1'
-            implementation 'org.controlsfx:controlsfx:11.2.0'
+            implementation 'org.controlsfx:controlsfx:11.2.1'
             implementation 'org.glassfish.jersey.core:jersey-client:2.29'
             implementation 'org.kordamp.ikonli:ikonli-javafx:12.3.1'
             implementation 'org.kordamp.ikonli:ikonli-material-pack:12.3.1'
@@ -96,7 +96,7 @@ subprojects {
             implementation 'org.slf4j:slf4j-jdk-platform-logging:2.0.12'
             implementation 'org.slf4j:jul-to-slf4j:2.0.12'
             implementation 'org.swinglabs.swingx:swingx-all:1.6.5-1'
-            runtimeOnly 'ch.qos.logback:logback-classic:1.4.14'
+            runtimeOnly 'ch.qos.logback:logback-classic:1.5.3'
             runtimeOnly 'org.fusesource.jansi:jansi:2.4.1'
         }
