+++ org.mbari.vars.ui/src/main/java/org/mbari/vars/ui/commands/ChangeConceptCmd.java
@@ -1,30 +1,29 @@
 package org.mbari.vars.ui.commands;
 
 import org.mbari.vars.services.model.Annotation;
+import org.mbari.vars.services.model.ObservationsUpdate;
+import org.mbari.vars.services.model.User;
+import org.mbari.vars.ui.UIToolBox;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
-/**
- * @author Brian Schlining
- * @since 2017-09-18T10:46:00
- * @deprecated use {@link ChangeConceptCmd2} instead
- */
-public class ChangeConceptCmd extends UpdateAnnotationsCmd {
-
-    private final String concept;
+public class ChangeConceptCmd extends UpdateObservationsCmd {
 
     public ChangeConceptCmd(List<Annotation> originalAnnotations, String concept) {
-        super(originalAnnotations, originalAnnotations.stream()
-                .map(Annotation::new)
-                .peek(a -> a.setConcept(concept))
-                .collect(Collectors.toList()), true, true);
-        this.concept = concept;
+        super(originalAnnotations, ObservationsUpdate.forConcept(observationUuids(originalAnnotations), concept));
+    }
+
+    @Override
+    public void apply(UIToolBox toolBox) {
+        final User user = toolBox.getData().getUser();
+        if (user != null) {
+            observationsUpdate = observationsUpdate.withObserver(user.getUsername());
+        }
+        super.apply(toolBox);
     }
 
     @Override
     public String getDescription() {
-        return "Changing concept for " + originalAnnotations.size() + " annotations to " + concept;
+        return "Changing concept of " + originalAnnotations.size() + " annotations to " + observationsUpdate.concept();
     }
-
 }
